requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Django==5.2.4
djangorestframework==3.16.0
dlib==20.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fonttools==4.59.0
gTTS==2.5.4
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pyttsx3==2.99
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
